// frontend/src/lib/generated/api/models/DTOPago.ts
export interface DTOPago {
    /**
     * 
     * @type {number}
     * @memberof DTOPago
     */
    id?: number;
    /**
     * 
     * @type {Date}
     * @memberof DTOPago
     */
    fechaPago?: Date;
    /**
     * 
     * @type {number}
     * @memberof DTOPago
     */
    importe?: number;
    /**
     * 
     * @type {string}
     * @memberof DTOPago
     */
    metodoPago?: DTOPagoMetodoPagoEnum;
    /**
     * 
     * @type {string}
     * @memberof DTOPago
     */
    estado?: DTOPagoEstadoEnum;
    /**
     * 
     * @type {string}
     * @memberof DTOPago
     */
    alumnoId?: string;
    /**
     * 
     * @type {boolean}
     * @memberof DTOPago
     */
    facturaCreada?: boolean;
    /**
     * 
     * @type {Array<ItemPago>}
     * @memberof DTOPago
     */
    items?: Array<ItemPago>;
    /**
     * 
     * @type {string}
     * @memberof DTOPago
     */
    stripePaymentIntentId?: string;
    /**
     * 
     * @type {string}
     * @memberof DTOPago
     */
    stripeChargeId?: string;
    /**
     * 
     * @type {string}
     * @memberof DTOPago
     */
    failureReason?: string;
    /**
     * 
     * @type {string}
     * @memberof DTOPago
     */
    clientSecret?: string;
    /**
     * 
     * @type {number}
     * @memberof DTOPago
     */
    classId?: number;
    /**
     * 
     * @type {string}
     * @memberof DTOPago
     */
    descripcionEstado?: string;
    /**
     * 
     * @type {number}
     * @memberof DTOPago
     */
    numeroItems?: number;
    /**
     * 
     * @type {string}
     * @memberof DTOPago
     */
    descripcionMetodoPago?: string;
}


/**
 * @export
 */
export const DTOPagoMetodoPagoEnum = {
    Efectivo: 'EFECTIVO',
    Debito: 'DEBITO',
    Credito: 'CREDITO',
    Transferencia: 'TRANSFERENCIA',
    Stripe: 'STRIPE'
} as const;
export type DTOPagoMetodoPagoEnum = typeof DTOPagoMetodoPagoEnum[keyof typeof DTOPagoMetodoPagoEnum];

/**
 * @export
 */
export const DTOPagoEstadoEnum = {
    Pendiente: 'PENDIENTE',
    Procesando: 'PROCESANDO',
    Exito: 'EXITO',
    Error: 'ERROR',
    Reembolsado: 'REEMBOLSADO'
} as const;
export type DTOPagoEstadoEnum = typeof DTOPagoEstadoEnum[keyof typeof DTOPagoEstadoEnum];