// frontend/src/lib/generated/api/models/DTOPeticionCrearPago.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Academia API
 * API para gestión de academia con roles de administrador, profesor y alumno
 *
 * The version of the OpenAPI document: 1.0
 * Contact: contact@academia.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Payment creation request
 * @export
 * @interface DTOPeticionCrearPago
 */
export interface DTOPeticionCrearPago {
    /**
     * 
     * @type {number}
     * @memberof DTOPeticionCrearPago
     */
    importe: number;
    /**
     * 
     * @type {string}
     * @memberof DTOPeticionCrearPago
     */
    alumnoId: string;
    /**
     * 
     * @type {string}
     * @memberof DTOPeticionCrearPago
     */
    description: string;
    /**
     * 
     * @type {string}
     * @memberof DTOPeticionCrearPago
     */
    currency: string;
    /**
     * 
     * @type {number}
     * @memberof DTOPeticionCrearPago
     */
    classId?: number;
}

/**
 * Check if a given object implements the DTOPeticionCrearPago interface.
 */
export function instanceOfDTOPeticionCrearPago(value: object): value is DTOPeticionCrearPago {
    if (!('importe' in value) || value['importe'] === undefined) return false;
    if (!('alumnoId' in value) || value['alumnoId'] === undefined) return false;
    if (!('description' in value) || value['description'] === undefined) return false;
    if (!('currency' in value) || value['currency'] === undefined) return false;
    return true;
}

export function DTOPeticionCrearPagoFromJSON(json: any): DTOPeticionCrearPago {
    return DTOPeticionCrearPagoFromJSONTyped(json, false);
}

export function DTOPeticionCrearPagoFromJSONTyped(json: any, ignoreDiscriminator: boolean): DTOPeticionCrearPago {
    if (json == null) {
        return json;
    }
    return {
        
        'importe': json['importe'],
        'alumnoId': json['alumnoId'],
        'description': json['description'],
        'currency': json['currency'],
        'classId': json['classId'] == null ? undefined : json['classId'],
    };
}

export function DTOPeticionCrearPagoToJSON(json: any): DTOPeticionCrearPago {
    return DTOPeticionCrearPagoToJSONTyped(json, false);
}

export function DTOPeticionCrearPagoToJSONTyped(value?: DTOPeticionCrearPago | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'importe': value['importe'],
        'alumnoId': value['alumnoId'],
        'description': value['description'],
        'currency': value['currency'],
        'classId': value['classId'],
    };
}