// frontend/src/lib/generated/api/models/DTORespuestaPaginada.ts
export interface DTORespuestaPaginada {
    /**
     * List of elements for the current page
     * @type {Array<any>}
     * @memberof DTORespuestaPaginada
     */
    content: Array<any>;
    /**
     * Current page number (0-indexed)
     * @type {number}
     * @memberof DTORespuestaPaginada
     */
    page: number;
    /**
     * Page size
     * @type {number}
     * @memberof DTORespuestaPaginada
     */
    size: number;
    /**
     * Total number of elements
     * @type {number}
     * @memberof DTORespuestaPaginada
     */
    totalElements: number;
    /**
     * Total number of pages
     * @type {number}
     * @memberof DTORespuestaPaginada
     */
    totalPages: number;
    /**
     * Indicates if this is the first page
     * @type {boolean}
     * @memberof DTORespuestaPaginada
     */
    isFirst: boolean;
    /**
     * Indicates if this is the last page
     * @type {boolean}
     * @memberof DTORespuestaPaginada
     */
    isLast: boolean;
    /**
     * Indicates if the page has content
     * @type {boolean}
     * @memberof DTORespuestaPaginada
     */
    hasContent: boolean;
    /**
     * Field by which results are sorted
     * @type {string}
     * @memberof DTORespuestaPaginada
     */
    sortBy: string;
    /**
     * Sort direction
     * @type {string}
     * @memberof DTORespuestaPaginada
     */
    sortDirection: string;
}